# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Veylora Global API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResult {
  error: string;
}

export interface MessageResult {
  message: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface LoginResult {
  success: boolean;
  username: string;
  role: string;
}

export interface SessionResult {
  authenticated: boolean;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  role?: string | null;
}

export interface TrackInput {
  /** @minLength 5 */
  consNo: string;
}

export interface Courier {
  id: number;
  consNo: string;
  sName: string;
  sMail: string;
  sPhone: string;
  sAdd: string;
  rName: string;
  rMail: string;
  rPhone: string;
  rAdd: string;
  type: string;
  weight: string;
  invoiceNo: string;
  qty: number;
  freight: string;
  mode: string;
  pmode: string;
  pickDate: string;
  deptDate: string;
  status: string;
  product: string;
  origin: string;
  destination: string;
  /** @nullable */
  lat?: string | null;
  /** @nullable */
  lon?: string | null;
}

export interface TrackUpdate {
  id: number;
  cid: number;
  consNo: string;
  updateDate: string;
  currentCity: string;
  newStatus: string;
  comments: string;
  currentLocation: string;
  bkTime: string;
  /** @nullable */
  rName?: string | null;
  /** @nullable */
  rPhone?: string | null;
  /** @nullable */
  rMail?: string | null;
  /** @nullable */
  rAdd?: string | null;
}

export interface TrackingResult {
  courier: Courier;
  updates: TrackUpdate[];
}

export interface ContactInput {
  name: string;
  email: string;
  subject: string;
  message: string;
}

export interface DashboardStats {
  totalShipments: number;
  inTransit: number;
  delivered: number;
  totalOfficers: number;
  totalOffices: number;
  recentShipments: Courier[];
}

export interface CourierInput {
  consNo: string;
  sName: string;
  sMail: string;
  sPhone: string;
  sAdd: string;
  rName: string;
  rMail: string;
  rPhone: string;
  rAdd: string;
  type: string;
  weight: string;
  invoiceNo: string;
  qty: number;
  freight: string;
  mode: string;
  pmode: string;
  pickDate: string;
  deptDate: string;
  status: string;
  product: string;
  origin: string;
  destination: string;
  lat?: string;
  lon?: string;
}

export interface CourierUpdate {
  sName?: string;
  sMail?: string;
  sPhone?: string;
  sAdd?: string;
  rName?: string;
  rMail?: string;
  rPhone?: string;
  rAdd?: string;
  type?: string;
  weight?: string;
  qty?: number;
  freight?: string;
  mode?: string;
  pmode?: string;
  pickDate?: string;
  deptDate?: string;
  status?: string;
  product?: string;
  origin?: string;
  destination?: string;
  lat?: string;
  lon?: string;
}

export interface TrackUpdateInput {
  updateDate: string;
  currentCity: string;
  newStatus: string;
  comments: string;
  currentLocation: string;
  bkTime: string;
  rName?: string;
  rPhone?: string;
  rMail?: string;
  rAdd?: string;
}

export interface Officer {
  id: number;
  officerName: string;
  address: string;
  email: string;
  phNo: string;
  office: string;
  regDate: string;
}

export interface OfficerInput {
  officerName: string;
  offPwd: string;
  address: string;
  email: string;
  phNo: string;
  office: string;
}

export interface Office {
  id: number;
  offName: string;
  address: string;
  city: string;
  phNo: string;
  officeTime: string;
  contactPerson: string;
}

export interface OfficeInput {
  offName: string;
  address: string;
  city: string;
  phNo: string;
  officeTime: string;
  contactPerson: string;
}