# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Veylora Global API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
...
  /couriers:
    get:
      operationId: listCouriers
      tags: [couriers]
      summary: List all shipments
      responses:
        "200":
          description: List of couriers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Courier"
    post:
      operationId: createCourier
      tags: [couriers]
      summary: Create a new shipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourierInput"
      responses:
        "201":
          description: Courier created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Courier"
  /couriers/{id}:
    get:
      operationId: getCourier
      tags: [couriers]
      summary: Get a shipment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Courier details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Courier"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
    put:
      operationId: updateCourier
      tags: [couriers]
      summary: Update a shipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourierUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Courier"
    delete:
      operationId: deleteCourier
      tags: [couriers]
      summary: Delete a shipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResult"

  # ─── Tracking Updates (Admin) ────────────────────────────────────────────
...
  /officers:
    get:
      operationId: listOfficers
      tags: [officers]
      summary: List all officers
      responses:
        "200":
          description: List of officers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Officer"
    post:
      operationId: createOfficer
      tags: [officers]
      summary: Create a new officer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfficerInput"
      responses:
        "201":
          description: Officer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Officer"
  /officers/{id}:
    delete:
      operationId: deleteOfficer
      tags: [officers]
      summary: Delete an officer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResult"

  # ─── Offices (Admin) ─────────────────────────────────────────────────────
  /offices:
    get:
      operationId: listOffices
      tags: [offices]
      summary: List all offices
      responses:
        "200":
          description: List of offices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Office"
    post:
      operationId: createOffice
      tags: [offices]
      summary: Create a new office
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfficeInput"
      responses:
        "201":
          description: Office created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Office"
  /offices/{id}:
    delete:
      operationId: deleteOffice
      tags: [offices]
      summary: Delete an office
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResult"
components:
  schemas:
...
    ErrorResult:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResult:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ─── Auth ───────────────────────────────────────────────────────────────
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    LoginResult:
      type: object
      required: [success, username, role]
      properties:
        success:
          type: boolean
        username:
          type: string
        role:
          type: string
...
    Courier:
      type: object
      required:
        - id
        - consNo
        - sName
        - sMail
        - sPhone
        - sAdd
        - rName
        - rMail
        - rPhone
        - rAdd
        - type
        - weight
        - invoiceNo
        - qty
        - freight
        - mode
        - pmode
        - pickDate
        - deptDate
        - status
        - product
        - origin
        - destination
      properties:
        id:
          type: integer
        consNo:
          type: string
        sName:
          type: string
        sMail:
          type: string
        sPhone:
          type: string
        sAdd:
          type: string
        rName:
          type: string
        rMail:
          type: string
        rPhone:
          type: string
        rAdd:
          type: string
        type:
          type: string
        weight:
          type: string
        invoiceNo:
          type: string
        qty:
          type: integer
        freight:
          type: string
        mode:
          type: string
        pmode:
          type: string
        pickDate:
          type: string
        deptDate:
          type: string
        status:
          type: string
        product:
          type: string
        origin:
          type: string
        destination:
          type: string
        lat:
          type: ["string", "null"]
        lon:
          type: ["string", "null"]
    CourierInput:
      type: object
      required:
        - consNo
        - sName
        - sMail
        - sPhone
        - sAdd
        - rName
        - rMail
        - rPhone
        - rAdd
        - type
        - weight
        - invoiceNo
        - qty
        - freight
        - mode
        - pmode
        - pickDate
        - deptDate
        - status
        - product
        - origin
        - destination
      properties:
        consNo:
          type: string
        sName:
          type: string
        sMail:
          type: string
        sPhone:
          type: string
        sAdd:
          type: string
        rName:
          type: string
        rMail:
          type: string
        rPhone:
          type: string
        rAdd:
          type: string
        type:
          type: string
        weight:
          type: string
        invoiceNo:
          type: string
        qty:
          type: integer
        freight:
          type: string
        mode:
          type: string
        pmode:
          type: string
        pickDate:
          type: string
        deptDate:
          type: string
        status:
          type: string
        product:
          type: string
        origin:
          type: string
        destination:
          type: string
        lat:
          type: string
        lon:
          type: string
    CourierUpdate:
      type: object
      properties:
        sName:
          type: string
        sMail:
          type: string
        sPhone:
          type: string
        sAdd:
          type: string
        rName:
          type: string
        rMail:
          type: string
        rPhone:
          type: string
        rAdd:
          type: string
        type:
          type: string
        weight:
          type: string
        qty:
          type: integer
        freight:
          type: string
        mode:
          type: string
        pmode:
          type: string
        pickDate:
          type: string
        deptDate:
          type: string
        status:
          type: string
        product:
          type: string
        origin:
          type: string
        destination:
          type: string
        lat:
          type: string
        lon:
          type: string

    # ─── Track Update ────────────────────────────────────────────────────────
...
    Officer:
      type: object
      required: [id, officerName, address, email, phNo, office, regDate]
      properties:
        id:
          type: integer
        officerName:
          type: string
        address:
          type: string
        email:
          type: string
        phNo:
          type: string
        office:
          type: string
        regDate:
          type: string
    OfficerInput:
      type: object
      required: [officerName, offPwd, address, email, phNo, office]
      properties:
        officerName:
          type: string
        offPwd:
          type: string
        address:
          type: string
        email:
          type: string
        phNo:
          type: string
        office:
          type: string

    # ─── Office ──────────────────────────────────────────────────────────────
    Office:
      type: object
      required: [id, offName, address, city, phNo, officeTime, contactPerson]
      properties:
        id:
          type: integer
        offName:
          type: string
        address:
          type: string
        city:
          type: string
        phNo:
          type: string
        officeTime:
          type: string
        contactPerson:
          type: string
    OfficeInput:
      type: object
      required: [offName, address, city, phNo, officeTime, contactPerson]
      properties:
        offName:
          type: string
        address:
          type: string
        city:
          type: string
        phNo:
          type: string
        officeTime:
          type: string
        contactPerson:
          type: string