# lib/api-zod/src/generated/types/courierUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Veylora Global API
 * OpenAPI spec version: 0.1.0
 */

export interface CourierUpdate {
  sName?: string;
  sMail?: string;
  sPhone?: string;
  sAdd?: string;
  rName?: string;
  rMail?: string;
  rPhone?: string;
  rAdd?: string;
  type?: string;
  weight?: string;
  qty?: number;
  freight?: string;
  mode?: string;
  pmode?: string;
  pickDate?: string;
  deptDate?: string;
  status?: string;
  product?: string;
  origin?: string;
  destination?: string;
  lat?: string;
  lon?: string;
}