# lib/api-zod/src/generated/types/office.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Veylora Global API
 * OpenAPI spec version: 0.1.0
 */

export interface Office {
  id: number;
  offName: string;
  address: string;
  city: string;
  phNo: string;
  officeTime: string;
  contactPerson: string;
}